+++ skforecast/ForecasterAutoregMultiSeriesCustom/tests/test_get_feature_importances.py
@@ -1,4 +1,4 @@
-# Unit test get_feature_importance ForecasterAutoregMultiSeriesCustom
+# Unit test get_feature_importances ForecasterAutoregMultiSeriesCustom
 # ==============================================================================
 import re
 import pytest
@@ -27,9 +27,9 @@ def create_predictors(y): # pragma: no cover
     return lags
                     
 
-def test_exception_is_raised_when_forecaster_is_not_fitted():
+def test_NotFittedError_is_raised_when_forecaster_is_not_fitted():
     """
-    Test exception is raised when calling get_feature_importance() and 
+    Test NotFittedError is raised when calling get_feature_importances() and 
     forecaster is not fitted.
     """
     forecaster = ForecasterAutoregMultiSeriesCustom(
@@ -40,15 +40,15 @@ def test_exception_is_raised_when_forecaster_is_not_fitted():
 
     err_msg = re.escape(
                 ("This forecaster is not fitted yet. Call `fit` with appropriate "
-                 "arguments before using `get_feature_importance()`.")
+                 "arguments before using `get_feature_importances()`.")
               )
     with pytest.raises(NotFittedError, match = err_msg):         
-        forecaster.get_feature_importance()
+        forecaster.get_feature_importances()
 
 
-def test_output_get_feature_importance_when_regressor_is_RandomForest():
+def test_output_get_feature_importances_when_regressor_is_RandomForest():
     """
-    Test output of get_feature_importance when regressor is RandomForestRegressor with lags=3
+    Test output of get_feature_importances when regressor is RandomForestRegressor with lags=3
     and it is trained with series pandas DataFrame.
     """
 
@@ -57,20 +57,20 @@ def test_output_get_feature_importance_when_regressor_is_RandomForest():
                      fun_predictors  = create_predictors,
                      window_size     = 3
                  )
-    
     forecaster.fit(series=series)
+
+    results = forecaster.get_feature_importances()
     expected = pd.DataFrame({
-                    'feature': ['custom_predictor_0', 'custom_predictor_1', 'custom_predictor_2', '1', '2'],
-                    'importance': np.array([0.9446366782006932, 0.0, 0.05536332179930687, 0.0, 0.0])
-                })
-    results = forecaster.get_feature_importance()
+                   'feature': ['custom_predictor_0', 'custom_predictor_1', 'custom_predictor_2', '1', '2'],
+                   'importance': np.array([0.9446366782006932, 0.0, 0.05536332179930687, 0.0, 0.0])
+               })
 
-    pd.testing.assert_frame_equal(expected, results)
+    pd.testing.assert_frame_equal(results, expected)
 
 
-def test_output_get_feature_importance_when_regressor_is_RandomForest_with_exog():
+def test_output_get_feature_importances_when_regressor_is_RandomForest_with_exog():
     """
-    Test output of get_feature_importance when regressor is RandomForestRegressor with lags=3
+    Test output of get_feature_importances when regressor is RandomForestRegressor with lags=3
     and it is trained with series pandas DataFrame and a exogenous variable
     exog=pd.Series(np.arange(10, 20), name='exog').
     """
@@ -80,18 +80,19 @@ def test_output_get_feature_importance_when_regressor_is_RandomForest_with_exog(
                      window_size     = 3
                  )
     forecaster.fit(series=series, exog=pd.Series(np.arange(10, 20), name='exog'))
+
+    results = forecaster.get_feature_importances()
     expected = pd.DataFrame({
-                    'feature': ['custom_predictor_0', 'custom_predictor_1', 'custom_predictor_2', 'exog', '1', '2'],
-                    'importance': np.array([0.73269896, 0., 0.21193772, 0.05536332, 0., 0.])
+                   'feature': ['custom_predictor_0', 'custom_predictor_1', 'custom_predictor_2', 'exog', '1', '2'],
+                   'importance': np.array([0.73269896, 0., 0.21193772, 0.05536332, 0., 0.])
                })
-    results = forecaster.get_feature_importance()
 
-    pd.testing.assert_frame_equal(expected, results)
+    pd.testing.assert_frame_equal(results, expected)
 
 
-def test_output_get_feature_importance_when_regressor_is_LinearRegression():
+def test_output_get_feature_importances_when_regressor_is_LinearRegression():
     """
-    Test output of get_feature_importance when regressor is LinearRegression with lags=3
+    Test output of get_feature_importances when regressor is LinearRegression with lags=3
     and it is trained with series pandas DataFrame.
     """
     forecaster = ForecasterAutoregMultiSeriesCustom(
@@ -100,66 +101,74 @@ def test_output_get_feature_importance_when_regressor_is_LinearRegression():
                      window_size     = 3
                  )
     forecaster.fit(series=series)
+
+    results = forecaster.get_feature_importances()
     expected = pd.DataFrame({
-                    'feature': ['custom_predictor_0', 'custom_predictor_1', 'custom_predictor_2', '1', '2'],
-                    'importance': np.array([3.33333333e-01, 3.33333333e-01, 3.33333333e-01, 
-                                            -1.48164367e-16, 1.48164367e-16])
+                   'feature': ['custom_predictor_0', 'custom_predictor_1', 'custom_predictor_2', '1', '2'],
+                   'importance': np.array([3.33333333e-01, 3.33333333e-01, 3.33333333e-01, 
+                                           -1.48164367e-16, 1.48164367e-16])
                })
-    results = forecaster.get_feature_importance()
 
-    pd.testing.assert_frame_equal(expected, results)
+    pd.testing.assert_frame_equal(results, expected)
 
 
-def test_output_get_feature_importance_when_regressor_is_LinearRegression_with_exog():
+def test_output_get_feature_importances_when_regressor_is_LinearRegression_with_exog():
     """
-    Test output of get_feature_importance when regressor is LinearRegression with lags=3
+    Test output of get_feature_importances when regressor is LinearRegression with lags=3
     and it is trained with series pandas DataFrame and a exogenous variable
     exog=pd.Series(np.arange(10, 15), name='exog').
     """
-    series = pd.DataFrame({'1': pd.Series(np.arange(5)), 
-                           '2': pd.Series(np.arange(5))
-                          })
-
+    series_2 = pd.DataFrame({'1': pd.Series(np.arange(5)), 
+                             '2': pd.Series(np.arange(5))})
     forecaster = ForecasterAutoregMultiSeriesCustom(
                      regressor       = LinearRegression(),
                      fun_predictors  = create_predictors,
                      window_size     = 3
                  )
-    forecaster.fit(series=series, exog=pd.Series(np.arange(10, 15), name='exog'))
+    forecaster.fit(series=series_2, exog=pd.Series(np.arange(10, 15), name='exog'))
+
+    results = forecaster.get_feature_importances()
     expected = pd.DataFrame({
-                    'feature': ['custom_predictor_0', 'custom_predictor_1', 'custom_predictor_2', 'exog', '1', '2'],
-                    'importance': np.array([2.50000000e-01, 2.50000000e-01, 2.50000000e-01, 
-                                            2.50000000e-01, -8.32667268e-17, 8.32667268e-17])
+                   'feature': ['custom_predictor_0', 'custom_predictor_1', 'custom_predictor_2', 'exog', '1', '2'],
+                   'importance': np.array([2.50000000e-01, 2.50000000e-01, 2.50000000e-01, 
+                                           2.50000000e-01, -8.32667268e-17, 8.32667268e-17])
                })
-    results = forecaster.get_feature_importance()
 
-    pd.testing.assert_frame_equal(expected, results)
+    pd.testing.assert_frame_equal(results, expected)
 
 
-def test_output_get_feature_importance_when_regressor_no_attributes():
+def test_output_and_UserWarning_get_feature_importances_when_regressor_no_attributes():
     """
-    Test output of get_feature_importance when regressor is MLPRegressor with lags=3
+    Test output of get_feature_importances when regressor is MLPRegressor with lags=3
     and it is trained with series pandas DataFrame. Since MLPRegressor hasn't attributes
-    `feature_importances_` or `coef_, results = None and a warning is raised`
+    `feature_importances_` or `coef_, results = None and a UserWarning is issues.
     """
-    series = pd.DataFrame({'1': pd.Series(np.arange(5)), 
-                           '2': pd.Series(np.arange(5))
-                          })
+    series_2 = pd.DataFrame({'1': pd.Series(np.arange(5)), 
+                             '2': pd.Series(np.arange(5))})
     forecaster = ForecasterAutoregMultiSeriesCustom(
                      regressor       = MLPRegressor(solver = 'lbfgs', max_iter= 50, random_state=123),
                      fun_predictors  = create_predictors,
                      window_size     = 3
                  )
-    forecaster.fit(series=series)
+    forecaster.fit(series=series_2)
+
+    estimator = forecaster.regressor
     expected = None
-    results = forecaster.get_feature_importance()
 
-    assert results is expected
+    warn_msg = re.escape(
+                (f"Impossible to access feature importances for regressor of type "
+                 f"{type(estimator)}. This method is only valid when the "
+                 f"regressor stores internally the feature importances in the "
+                 f"attribute `feature_importances_` or `coef_`.")
+            )
+    with pytest.warns(UserWarning, match = warn_msg):
+        results = forecaster.get_feature_importances()
+        assert results is expected
 
 
-def test_output_get_feature_importance_when_pipeline():
+def test_output_get_feature_importances_when_pipeline_LinearRegression():
     """
-    Test output of get_feature_importance when regressor is pipeline,
+    Test output of get_feature_importances when regressor is pipeline,
     (StandardScaler() + LinearRegression with lags=3),
     it is trained with series pandas DataFrame.
     """  
@@ -169,11 +178,35 @@ def test_output_get_feature_importance_when_pipeline():
                      window_size     = 3
                  )
     forecaster.fit(series=series)
-    expected = pd.DataFrame({
-                    'feature': ['custom_predictor_0', 'custom_predictor_1', 'custom_predictor_2', '1', '2'],
-                    'importance': np.array([6.66666667e-01, 6.66666667e-01, 6.66666667e-01, 
-                                            -7.40821837e-17, 7.40821837e-17])
-               })
-    results = forecaster.get_feature_importance()
 
-    pd.testing.assert_frame_equal(expected, results)
+    results = forecaster.get_feature_importances()
+    expected = pd.DataFrame({
+                   'feature': ['custom_predictor_0', 'custom_predictor_1', 'custom_predictor_2', '1', '2'],
+                   'importance': np.array([6.66666667e-01, 6.66666667e-01, 6.66666667e-01, 
+                                           -7.40821837e-17, 7.40821837e-17])
+               })
+
+    pd.testing.assert_frame_equal(results, expected)
+
+
+def test_output_get_feature_importances_when_pipeline_RandomForestRegressor():
+    """
+    Test output of get_feature_importances when regressor is pipeline,
+    (StandardScaler() + RandomForestRegressor with lags=3),
+    it is trained with series pandas DataFrame.
+    """
+    forecaster = ForecasterAutoregMultiSeriesCustom(
+                     regressor = make_pipeline(StandardScaler(), 
+                                               RandomForestRegressor(n_estimators=1, max_depth=2, random_state=123)),
+                     fun_predictors  = create_predictors,
+                     window_size     = 3
+                 )
+    forecaster.fit(series=series)
+
+    results = forecaster.get_feature_importances()
+    expected = pd.DataFrame({
+                   'feature': ['custom_predictor_0', 'custom_predictor_1', 'custom_predictor_2', '1', '2'],
+                   'importance': np.array([0.9446366782006932, 0.0, 0.05536332179930687, 0.0, 0.0])
+               })
+
+    pd.testing.assert_frame_equal(results, expected)