+++ skforecast/ForecasterSarimax/tests/test_get_feature_importances.py
@@ -1,4 +1,4 @@
-# Unit test get_feature_importance ForecasterSarimax
+# Unit test get_feature_importances ForecasterSarimax
 # ==============================================================================
 import re
 import pytest
@@ -11,22 +11,22 @@ from sklearn.exceptions import NotFittedError
 
 def test_exception_is_raised_when_forecaster_is_not_fitted():
     """
-    Test exception is raised when calling get_feature_importance() and 
+    Test exception is raised when calling get_feature_importances() and 
     forecaster is not fitted.
     """
     forecaster = ForecasterSarimax(regressor=ARIMA(order=(1,1,1)))
 
     err_msg = re.escape(
                 ("This forecaster is not fitted yet. Call `fit` with appropriate "
-                 "arguments before using `get_feature_importance()`.")
+                 "arguments before using `get_feature_importances()`.")
               )
     with pytest.raises(NotFittedError, match=err_msg):         
-        forecaster.get_feature_importance()
+        forecaster.get_feature_importances()
 
 
-def test_output_get_feature_importance_ForecasterSarimax():
+def test_output_get_feature_importances_ForecasterSarimax():
     """
-    Test output of get_feature_importance ForecasterSarimax.
+    Test output of get_feature_importances ForecasterSarimax.
     """
     forecaster = ForecasterSarimax(regressor=ARIMA(order=(1,1,1)))
     forecaster.fit(y=pd.Series(np.arange(10)))
@@ -35,6 +35,6 @@ def test_output_get_feature_importance_ForecasterSarimax():
                     'importance': np.array([0.49998574676910396, 0.5000130662306124, 
                                             7.479723906909597e-11, 2.658043128694438e-12])
                 })
-    results = forecaster.get_feature_importance()
+    results = forecaster.get_feature_importances()
 
     pd.testing.assert_frame_equal(expected, results)