+++ skforecast/__init__.py
@@ -1,2 +1,2 @@
 name = "skforecast"
-__version__ = "0.7.1"
+__version__ = "0.8.0"